requirements.txt
hdx-python-api==4.6.1
PyMySQL==0.10.0
python-slugify==4.0.1
